backend/pyproject.toml
[project]
name = "domain-blacklist-checker"
version = "0.1.0"
description = "Domain Blacklist & Spam Checker backend"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.21.0",
  "dnspython>=2.7.0",
  "fastapi>=0.116.0",
  "greenlet>=3.1.1",
  "httpx[socks]>=0.28.0",
  "python-multipart>=0.0.20",
  "pydantic-settings>=2.8.0",
  "sqlalchemy[asyncio]>=2.0.38",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.9.10",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
